requirements.txt
jupyterlab>3.0

langchain-nvidia-ai-endpoints
faiss-cpu
faiss-gpu
pypdf2
bs4
langchain
langchain-community
jupyterlab_widgets
ipywidgets
transformers
langchain_huggingface
scikit-learn
umap-learn
selenium
sentence_transformers
einops
opt_einsum
bokeh
pypdf
tqdm
webdriver-manager
